+++ src/util/validateConfig.js
@@ -1,11 +1,18 @@
 import log from './log'
 
 export function validateConfig(config) {
-  if (config.content.files.length === 0) {
+  if (config.content.files !== 'auto' && config.content.files.length === 0) {
     log.warn('content-problems', [
       'The `content` option in your Tailwind CSS configuration is missing or empty.',
       'Configure your content sources or your generated CSS will be missing styles.',
       'https://tailwindcss.com/docs/content-configuration',
+    ])
+  }
+
+  if (config.content.files === 'auto') {
+    log.warn('auto-content-experimental', [
+      'Automatic content detection in Tailwind CSS is currently in experimental preview.',
+      'Preview features are not covered by semver, and may change at any time.',
     ])
   }